WIP - fix perf issues
